@@ -21,7 +21,7 @@ include_package_data = True
 package_dir =
     = src
 packages = find:
-python_requires = >=3.6
+python_requires = >=3.7
 install_requires =
     greenlet
 
