@@ -4,8 +4,7 @@ import unittest
 from greenletio.core import bridge, async_
 from greenletio.green import socket
 
-if not hasattr(asyncio, 'create_task'):
-    asyncio.create_task = asyncio.ensure_future
+from .util import run_coro
 
 
 class TestSocket(unittest.TestCase):
@@ -46,10 +45,7 @@ class TestSocket(unittest.TestCase):
             while var is None:
                 await asyncio.sleep(0)
 
-        if sys.platform == 'win32':
-            loop = asyncio.SelectorEventLoop()
-            asyncio.set_event_loop(loop)
-        asyncio.get_event_loop().run_until_complete(main())
+        run_coro(main())
         assert var == b'HELLO'
 
     def test_sendto_recvfrom(self):
@@ -79,8 +75,5 @@ class TestSocket(unittest.TestCase):
             while var is None:
                 await asyncio.sleep(0)
 
-        if sys.platform == 'win32':
-            loop = asyncio.SelectorEventLoop()
-            asyncio.set_event_loop(loop)
-        asyncio.get_event_loop().run_until_complete(main())
+        run_coro(main())
         assert var == b'HELLO'
